ADR-002 stage 2.7 (1/2): split register_settings across 4 page modules
register_settings had grown to 901 lines, well past the ADR's 300-line
registration-function cap that stage 2.6 set but never actually enforced
for this function. Splits it into one register_settings_*_intents per
Settings page (general, api_provider, ollama, llama_cpp) under
backend/api/, following the same pattern stage 2.6 used for
register_canvas. register_settings itself is now a ~25-line orchestrator.

apply_ollama_reasoning_level and its 5 siblings, plus the renamed
run_locked/locked_llama_cpp_settings/republish_composer_reasoning/
redact_secret helpers, move to a new leaf module
(backend/api/_settings_shared.py) rather than staying in backend/settings.py:
backend.composer imports several of them directly, and the new intent
modules need to call them too, so leaving them in backend/settings.py
would have made backend/settings.py and the new intent modules import
each other. backend/settings.py re-exports the six apply_* functions and
run_locked so composer.py's existing imports need only a name change,
not a new import path.

Verification surfaced two real bugs beyond pure code movement:
composer.py had its own private `_apply` cross-import that an initial
grep for dotted-attribute usage missed (it was a `from x import y`
form), and 3 tests in test_settings.py monkeypatched
backend.settings._apply directly - which stopped reaching the code under
test once each intent module held its own independent `run_locked` name
binding. Both are fixed; full suite (1236 tests) passes.

register_chat_library (319 lines) and the AST test enforcing the
300-line cap repo-wide are the remaining stage 2.7 work, in a follow-up
PR.
